### 2025-07-01 #8524 @editorialbot
[PRE REVIEW]: scoup: Simulate Codon Sequences with Darwinian Selection Incorporated as an Ornstein-Uhlenbeck Process
**Submitting author:** @thsadiq (<a href="http://orcid.org/0000-0003-0192-7134">Hassan Sadiq</a>)
**Repository:** https://github.com/thsadiq/scoup
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.3.0
