### 2025-07-02 eecb02c
Create CONTRIBUTING.md
