+++ CONTRIBUTING.md
@@ -1,0 +1,26 @@
+# Introduction
+Thank you for considering contributing to the `scoup` codon simulation
+R/Bioconductor package. Your input will definitely be treated with great
+importance.
+
+# Proposing an update
+Should you wish to propose an update to the package, kindly raise an issue
+on the dedicated [Github page](https://github.com/thsadiq/scoup/issues).
+Rest assured that your proposal will be seriously considered and possibly
+committed to the primary Bioconductor version.
+
+# Reporting issues or problems with the package
+Bugs and/or issues may be reported by creating an issue on the package's
+[Github page](https://github.com/thsadiq/scoup/issues).
+
+# Seeking support
+Users that need assistance using the package or any of the functions
+therein should please consult the function description file embedded
+with the package. Detailed examples are also available on the package's
+[vignette page](https://bioconductor.org/packages/release/bioc/vignettes/scoup/inst/doc/scoup.html)
+that is hosted on the Bioconductor platform. Further assistance may be
+sought by posting on the
+[Bioconductor Forum](https://support.bioconductor.org/) with the `scoup`
+tag. Alternatively, users are able to send an e-mail directly to the
+corresponding author, [Hassan Sadiq](hassan.t.sadiq@gmail.com) or to
+the very helpful [Bioconductor Mailing List](bioc-devel@r-project.org).
