### 2025-07-03 be95394
Update README.md for joss submission
