### 2025-07-03 7c5d378
Update README.md for JOSS submission
