### 2025-08-04 #8524 @arfon
@thsadiq – would you mind merging your `main` branch into the `joss` branch so we can run our summary statistics on the software for this submission please?

Also, we're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
