+++ joss/README.md
@@ -1,0 +1,2 @@
+# Folder Information
+This folder contains files necessary to facilitate submission of an article that describes the `scoup` package to the Journal of Open Source Software.
