+++ README.md
@@ -1,2 +1,0 @@
-# Branch Information
-This branch contains files necessary to facilitate submission of an article that describes the `scoup` package to the Journal of Open Source Software.
