+++ paper/README.md
@@ -1,2 +1,6 @@
+<<<<<<< HEAD
 # Folder Information
 This folder contains files necessary to facilitate submission of an article that describes the `scoup` package to the Journal of Open Source Software.
+=======
+Hello
+>>>>>>> ca981b9 (Named joss folder appropriately)
