### 2025-08-05 5ea2a4a
Update README.md
JOSS oupblication update
