+++ paper/README.md
@@ -1,6 +1,11 @@
 <<<<<<< HEAD
+<<<<<<< HEAD
 # Folder Information
 This folder contains files necessary to facilitate submission of an article that describes the `scoup` package to the Journal of Open Source Software.
 =======
 Hello
 >>>>>>> ca981b9 (Named joss folder appropriately)
+=======
+# Folder Information
+This folder contains files necessary to facilitate submission of an article that describes the `scoup` package to the Journal of Open Source Software.
+>>>>>>> ab173a0 (Update README.md)
