+++ paper/paper.bib
@@ -1,0 +1,136 @@
+% ><>< ========================================================================= ><>< %
+% ><><   scoup: Simulate Codon Sequences with Darwinian Selection Incorporated   ><>< %
+% ><><                     as an Ornstein-Uhlenbeck Process.                     ><>< %
+% ><><                           ~~~~~~~~~~~~~~~~~~~~~                           ><>< %
+% ><><                               Bibliography.                               ><>< %
+% ><><                           ~~~~~~~~~~~~~~~~~~~~~                           ><>< %
+% ><><                              13th June 2025.                              ><>< %
+% ><>< ========================================================================= ><>< %
+
+@article{jacques2023,
+  title   = "{Roadmap to the Study of Gene and Protein Phylogeny and Evolution
+            -- A Practical Guide}",
+  author  = {Florian Jacques and Paulina Bolivar and Kristian Pietras and
+            Emma U Hammarlund},
+  journal = "{PLoS One}",
+  year    = {2023},
+  pages   = {e0279597},
+  volume  = {18},
+  number  = {2},
+  doi     = {10.1371/journal.pone.0279597}
+}
+
+@article{yang2007,
+  title   = "{PAML 4: Phylogenetic Analysis by Maximum Likelihood}",
+  author  = {Ziheng Yang},
+  journal = {Molecular Biology and Evolution},
+  year    = {2007},
+  volume  = {24},
+  number  = {8},
+  pages   = {1586--1591},
+  doi     = {10.1093/molbev/msm088}
+}
+
+@article{hyphy2020,
+  title   = "{{HyPhy 2.5} -- A Customizable Platform for Evolutionary Hypothesis
+              Testing Using Phylogenies}",
+  author  = {Sergei L {Kosakovsky Pond} and Art F Y Poon and Ryan Velazquez and
+             Steven Weaver and N Lance Hepler and Ben Murrell and Stephen D Shank
+             and Brittany Rife Magalis and Dave Bouvier and Anton Nekrutenko and
+             Sadie Wisotsky and Stephanie J Spielman and Simon D W Frost
+             and S V Muse},
+  journal = {Molecular Biology and Evolution},
+  year    = {2020},
+  pages   = {295--299},
+  volume  = {37},
+  number  = {1},
+  doi     = {10.1093/molbev/msz197}
+}
+
+@article{bioc2004,
+  title  = "{Bioconductor: Open Software Development for Computational Biology
+             and Bioinformatics}",
+  author  = {Robert C Gentleman and Vincent J Carey and Douglas M Bates and
+             Ben Bolstad and Marcel Dettling and Sandrine Dudoit and Byron Ellis
+             and Laurent Gautier and Yongchao Ge and Jeff Gentry and
+             Kurt Hornik and Torsten Hothorn and Wolfgang Huber and Stefano Iacus
+             and Rafael Irizarry and Friedrich Leisch and Cheng Li and
+             Martin Maechler and Anthony J Rossini and Gunther Sawitzki and
+             Colin Smith and Gordon Smyth and Luke Tierney and Jean Y H Yang and
+             Jianhua Zhang},
+  journal = {Genome Biology},
+  year    = {2004},
+  pages   = {R80},
+  volume  = {5},
+  number  = {10},
+  doi     = {10.1186/gb-2004-5-10-r80}
+}
+
+@article{halpern1998,
+  title   = "{Evolutionary Distances for Protein-Coding Sequences: Modelling
+            Site-Specific Residue Frequencies}",
[…]
