+++ paper/paper.md
@@ -1,0 +1,190 @@
+---
+title: 'scoup: Simulate Codon Sequences with Darwinian
+        Selection Incorporated as an Ornstein-Uhlenbeck Process'
+tags:
+  - R
+  - Molecular Evolution
+  - Ornstein-Uhlenbeck
+  - Phylogenetics
+  - Population Genetics
+  - Simulation
+  - Statistics
+authors:
+  - name: Hassan Sadiq
+    corresponding: true
+    orcid: 0000-0003-0192-7134
+    affiliation: "1, 2"
+  - name: Darren P. Martin
+    orcid: 0000-0002-8785-0870
+    affiliation: 2
+affiliations:
+ - name: Department of Statistics and Actuarial Science,
+         Stellenbosch University, South Africa
+   index: 1
+ - name: Institute of Infectious Diseases and Molecular Medicine,
+         Division of Computational Biology,
+         Department of Integrative Biomedical Sciences,
+         University of Cape Town, South Africa
+   index: 2
+date: 13 June 2025
+bibliography: paper.bib
+---
+
+# Summary
+
+Genetic analyses of natural selection within and between populations have
+increasingly developed along separate paths. The two important genres of
+evolutionary biology (i.e. phylogenetics and population genetics) borne
+from the split can only benefit from research that seeks to bridge
+the gap. Simulation algorithms that combine fundamental concepts from
+both genres are important to achieve such unifying objective. 
+We introduce `scoup`, a codon sequence simulator that is implemented
+in R and hosted on the Bioconductor platform. There is hardly any other
+simulator dedicated to genetic sequence generation for natural selection
+analyses on the platform. Concepts from the Halpern-Bruno mutation-selection
+model and the Ornstein-Uhlenbeck (OU) evolutionary algorithm were creatively
+fused such that the end-product is a novelty with respect to computational
+genetic simulation. Users are able to seamlessly adjust the model parameters
+to mimic complex evolutionary procedures that may have been otherwise
+infeasible. For example, it is possible to explicitly interrogate the
+concepts of static and changing fitness landscapes with regards to Darwinian
+natural selection in the context of codon sequences from multiple
+populations.
+
+# Statement of need
+
+Statistical inference of the extent to which Darwinian natural selection has
+impacted genetic data commands a healthy portion of the phylogenetic
+literature [@jacques2023]. Validation of these largely codon-based models
+relies heavily on simulated data. Given the ever increasing diversity of
+natural selection inference models that exist [@yang2007; @hyphy2020],
+there is a need for more sophisticated simulators to match the
+expanding model complexities.
+
+Bioconductor [@bioc2004] is a leading platform where peer-reviewed
+bioinformatic software useful for biological data analyses are hosted. A
+search of the entries on the platform, in Version 3.19 on 29 October 2024,
+with keywords including, `codon`, `mutation`, `selection`, `simulate`, and
+`simulation` returned a total of 72 unique packages out of the 2300 available.
+None of the retrieved entries was dedicated to codon data simulation for
+natural selection analyses. Thus, `scoup` is designed on the basis of the
+mutation-selection (MutSel) framework [@halpern1998] as an overdue
+contribution to the void. Software and/or packages for simulating genetic
+sequences are also rare in the scientific literature [@gearty2024]. 
+
+# Algorithm
+
+`scoup` is further unique for at least three reasons. First, it incorporates
+Darwinian natural selection into the MutSel model in terms of variability of
[…]
