+++ paper/paper.md
@@ -103,11 +103,19 @@ sequence alignments made up of $1000$ codon sites and $8$ extant taxa for
 each of the parameter combinations presented in \autoref{testfig}. The
 phylogeny used was balanced and the length of its branches were $0.10$ each.
 The stochastic OU framework was implemented and other function inputs were
+<<<<<<< HEAD
 left at their default values (\autoref{pseudocode}). Analytical estimates of
 $\mathrm{d}N/\mathrm{d}S$ were obtained following @spielman2015 and were
 averaged over all selection coefficient updates at each site and across
 the alignment. Likelihood inferences of $\omega$ were obtained with `CODEML`
 in `PAML` [@yang2007].
+=======
+left at their default values (\autoref{pseudocode}). Estimates of
+$\mathrm{d}N/\mathrm{d}S$ were obtained following @spielman2015 and were
+averaged over all selection coefficient updates at each site and across
+the alignment. Inferences of $\omega$ were obtained with `CODEML` in
+`PAML` [@yang2007].
+>>>>>>> 31e1ba6 (Updated subfolder for joss submission)
 
 
 ![\label{pseudocode}**An example R code for simulating a codon sequence
@@ -127,6 +135,7 @@ in `PAML` [@yang2007].
   obtained homogeneously from each alignment generated for every combination
   of the stochastic landscape ($\Sigma^{2}_{}$ and $\theta$) and the
   Darwinian selection ($\sigma^{2}_{n}$ and $\sigma^{2}_{s}$) parameters.
+<<<<<<< HEAD
   The solid points represent the average analytical $\mathrm{d}N/\mathrm{d}S$
   estimates while the open squares represent the average likelihood $\omega$
   estimates, across $20$ independent codon sequence alignments. The widths of
@@ -144,6 +153,25 @@ approximately $0.9971$ was obtained when the $\omega$ and
 $\mathrm{d}N/\mathrm{d}S$ averages were compared. The standard errors
 ranged between $[0.0000,\;0.0077)$ and $(0.0004,\;0.0627)$ for the
 $\mathrm{d}N/\mathrm{d}S$ and $\omega$ estimates respectively. These
+=======
+  The filled circles represent the average $\mathrm{d}N/\mathrm{d}S$
+  estimates while the empty squares represent the average $\omega$ estimates,
+  across $20$ independent codon sequence alignments. The widths of the arrows
+  correspond to twice the standard errors. The dashed lines highlight point
+  of neutral selection effect.](FIG2.pdf)
+
+
+Estimates of $\omega$ and $\mathrm{d}N/\mathrm{d}S$ summarised in
+\autoref{testfig} strongly agree, except for the case of
+$(\sigma^{2}_{n},\sigma^{2}_{s})=(0.10,0.02)$. The suppressed $\omega$
+estimates, that is most pronounced for $\sigma^{2}_{n},\sigma^{2}_{s}>0$,
+is likely a consequence of the well-documented conservative property of
+homogeneous $\omega$ inference techniques (see for example, @nielsen1998).
+Regardless, a correlation coefficient of approximately $0.9971$ was obtained
+when the $\omega$ and $\mathrm{d}N/\mathrm{d}S$ averages were compared. The
+standard errors ranged between $[0.0000,\;0.0077)$ and $(0.0004,\;0.0627)$
+for the $\mathrm{d}N/\mathrm{d}S$ and $\omega$ estimates respectively. These
+>>>>>>> 31e1ba6 (Updated subfolder for joss submission)
 measures confirm that the outputs from `scoup` are accurate.
 
 # Conclusions
[…]
