+++ paper/.github/workflows/draft-pdf.yml
@@ -1,9 +1,8 @@
 name: Draft PDF
 on:
   push:
-    workflow_dispatch: {}
     branches:
-      - 'joss'
+      - joss
     paths:
       - paper.md
       - paper.bib
