+++ .github/workflows/draft.yml
@@ -1,6 +1,8 @@
 name: Draft PDF
 on:
   push:
+    branches:
+      - joss
     paths:
       - paper.md
       - paper.bib
