### 2025-08-05 682a77b
Update paper.md
