+++ .github/workflows/draft.yml
@@ -1,8 +1,6 @@
 name: Draft PDF
 on:
   push:
-    branches:
-      - joss
     paths:
       - paper.md
       - paper.bib
