### 2025-12-09 #8524 @jromanowska
Wonderful! @clauswilke, I still need to find a second reviewer, so in the meantime, you can look around [in the repo](https://github.com/thsadiq/scoup) or simply wait.
