### 2025-12-12 #8524 @jromanowska
:wave: @Euan-Furness, @emilydolson, @ayush9pandey, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
