### 2026-01-11 #8524 @jromanowska
@arfon - I've discussed this with @rkurchin (thanks for checking!) and this submission needs to be checked against the old requirements. Could you help with this? Should I first start the real review?
