### 2026-01-17 #9797 @jromanowska
@arfon, this needs to be handled under the old policy - e.g., above, the automatic check of the paper is wrong.
