### 2026-01-17 #9797 @clauswilke
- [x] **Repository:** Is the source code for this software available here [https://github.com/thsadiq/scoup](https://github.com/thsadiq/scoup)?
