### 2026-01-17 #9797 @clauswilke
@thsadiq Looks like your manuscript is missing required sections Software Design, Research Impact Statement, and AI usage disclosure (see last block of my checklist).
