### 2026-01-17 #9797 @jromanowska
@clauswilke, thank you for starting the review so fast. I'm afraid the checklist is not correct. There is a new policy that we have just rolled out (on Jan.5), however this review needs to be checked with regard to the older policy, since it was accepted to be reviewed before the new policy rollout.
Thus, [this is the required format](https://joss.readthedocs.io/en/pre-2026-submission/paper.html) and [this is the checklist](https://joss.readthedocs.io/en/pre-2026-submission/review_checklist.html). I'm sorry for this misunderstanding, @clauswilke. @thsadiq, you don't need to update the paper with these sections.

The EiCs are in the process of changing the links and checklists where they should be rolled back, but there is a lot of manual work, so this might take some days.
@arfon , @rkurchin - could you change the checklists in this thread?
