### 2026-01-18 #9797 @ogoeli
- [x] **Repository:** Is the source code for this software available here [https://github.com/thsadiq/scoup](https://github.com/thsadiq/scoup)?
