### 2026-01-18 #9797 @jromanowska
@ogoeli, please, read the [post above](https://github.com/openjournals/joss-reviews/issues/9797#issuecomment-3764335937) before you start reviewing this submission. The checklists here will be updated ASAP.
