### 2026-01-30 #9797 @clauswilke
@thsadiq I'm basically fine with your manuscript, with one exception that should be addressed. I don't know how you generated Fig. 3 in your manuscript. I don't think the relevant code or data are deposited anywhere. I also don't fully understand what is shown. Are the solid points calculated analytically using the Spielman & Wilke formula and the open points are estimates from simulated alignments? If yes, we expect them to be similar, which they mostly are, but not for the top row. Why is that? I think you need to comment on it.
