### 2026-01-30 #9797 @jromanowska
@ogoeli, I think you've generated a new checklist by mistake - I will need to remove that one and let's focus on the [first one](https://github.com/openjournals/joss-reviews/issues/9797#issuecomment-3765399522), which is adapted to the old policy.

I can see that you've checked all the points there - are you finished with the review? Please, let us know what comments or questions do you have for the authors.
