### 2026-01-30 #9797 @thsadiq
@clauswilke Thank you so much for your insightful feedback. Here are my responses.

- The relevant code used to generate Figure 3 is bundled with the package and available in the vignettes/extdata/ folder. I attempted to mention this under the Code Availability section with the "... detailed documentation and tutorial files." comment. I can make that statement more explicit if necessary.
- Yes, the solid points were calculated analytically and the open ones from the simulated alignments. I commented on the top row of the figure in Line 57. We believe that it is probably a consequence of the well-documented conservative property of the homogeneous omega inference techniques.

Please let me know if these address your concerns. Thanks again for the constructive and prompt feedback.
