### 2026-01-30 #9797 @clauswilke
@thsadiq Did you make any changes to the manuscript? I only see the version from two weeks ago. I think you need to make some changes.

- I now see your text in lines 55-57, but honestly it's pretty cryptic. Please try to say these things more explicitly, using simple language: "Solid points are calculated analytically, open points are inferred from simulations, we expect open points to match solid points. They don't do so for the top row, this is most likely due to ..." etc. (I don't mean you should use this text verbatim, just describing the type of language I would like to see.)

- Similarly, you need to state explicitly that the code for Figure 3 is in vignettes/extdata/. I would never have found this. I will also point out that the material in vignettes/extdata/ doesn't seem complete. First, it refers to "the first figure" and "the second figure" and I don't know what those are. Are those Figure 1 and Figure 3 of the manuscript? Or Figure 2 and Figure 3? And also, for "the second figure" (I assume this is Figure 3) the code seems incomplete. There's no code that actually generates the figure. I also see no code to calculate the analytical dN/dS values. Finally, I think you should provide all PAML control files, not just a sample control file. Basically, the material provided is nowhere near complete and doesn't allow me to reproduce the results.
