### 2026-02-02 34d7b94
Review 1 update of paper.md
JOSS publication incorporated
