+++ paper/paper.md
@@ -104,6 +104,7 @@ each of the parameter combinations presented in \autoref{testfig}. The
 phylogeny used was balanced and the length of its branches were $0.10$ each.
 The stochastic OU framework was implemented and other function inputs were
 <<<<<<< HEAD
+<<<<<<< HEAD
 left at their default values (\autoref{pseudocode}). Analytical estimates of
 $\mathrm{d}N/\mathrm{d}S$ were obtained following @spielman2015 and were
 averaged over all selection coefficient updates at each site and across
@@ -116,6 +117,13 @@ averaged over all selection coefficient updates at each site and across
 the alignment. Inferences of $\omega$ were obtained with `CODEML` in
 `PAML` [@yang2007].
 >>>>>>> 31e1ba6 (Updated subfolder for joss submission)
+=======
+left at their default values (\autoref{pseudocode}). Analytical estimates of
+$\mathrm{d}N/\mathrm{d}S$ were obtained following @spielman2015 and were
+averaged over all selection coefficient updates at each site and across
+the alignment. Likelihood inferences of $\omega$ were obtained with `CODEML`
+in `PAML` [@yang2007].
+>>>>>>> 48b33c4 (Review 1 update of paper.md)
 
 
 ![\label{pseudocode}**An example R code for simulating a codon sequence
@@ -136,11 +144,15 @@ the alignment. Inferences of $\omega$ were obtained with `CODEML` in
   of the stochastic landscape ($\Sigma^{2}_{}$ and $\theta$) and the
   Darwinian selection ($\sigma^{2}_{n}$ and $\sigma^{2}_{s}$) parameters.
 <<<<<<< HEAD
+<<<<<<< HEAD
+=======
+>>>>>>> 48b33c4 (Review 1 update of paper.md)
   The solid points represent the average analytical $\mathrm{d}N/\mathrm{d}S$
   estimates while the open squares represent the average likelihood $\omega$
   estimates, across $20$ independent codon sequence alignments. The widths of
   the arrows correspond to twice the standard errors. The dashed lines
   highlight point of neutral selection effect.](FIG2.pdf)
+<<<<<<< HEAD
 
 
 Estimates of $\omega$ and $\mathrm{d}N/\mathrm{d}S$ summarised in
@@ -172,6 +184,20 @@ when the $\omega$ and $\mathrm{d}N/\mathrm{d}S$ averages were compared. The
 standard errors ranged between $[0.0000,\;0.0077)$ and $(0.0004,\;0.0627)$
 for the $\mathrm{d}N/\mathrm{d}S$ and $\omega$ estimates respectively. These
 >>>>>>> 31e1ba6 (Updated subfolder for joss submission)
+=======
+
+
+Estimates of $\omega$ and $\mathrm{d}N/\mathrm{d}S$ summarised in
+\autoref{testfig} strongly agree as expected, except for the case of
+$(\sigma^{2}_{n},\sigma^{2}_{s})=(0.10,0.02)$. The mismatch in the top
+row is most likely a consequence of the well-documented conservative
+property of homogeneous $\omega$ likelihood inference techniques (see
+for example, @nielsen1998). Regardless, a correlation coefficient of
+approximately $0.9971$ was obtained when the $\omega$ and
+$\mathrm{d}N/\mathrm{d}S$ averages were compared. The standard errors
+ranged between $[0.0000,\;0.0077)$ and $(0.0004,\;0.0627)$ for the
+$\mathrm{d}N/\mathrm{d}S$ and $\omega$ estimates respectively. These
+>>>>>>> 48b33c4 (Review 1 update of paper.md)
 measures confirm that the outputs from `scoup` are accurate.
 
 # Conclusions
[…]
