### 2026-02-02 7e0d7f7
Review 1 update of paper.md
