+++ vignettes/exData/imgB/summaries.R
@@ -1,0 +1,158 @@
+# ><>< ================================================================ ><>< #
+# ><><     scoup: Simulate Codon Sequences with Darwinian Selection     ><>< #
+# ><><          Incorporated  as an Ornstein-Uhlenbeck Process          ><>< #
+# ><>< ================================================================ ><>< #
+
+# ><>< # Clear memory
+rm(list=ls())
+
+# ><>< # Specify data directory and filenames
+pryDIR <- path.expand( file.path("~"))
+dndsEst <- read.csv( file.path(pryDIR, "dnds.csv"), header=F, row.names=1)
+omegaEst <- read.csv( file.path(pryDIR, "omega.csv") )
+imgfile <- file.path(pryDIR, "FIG3.pdf")
+
+# ><>< # Generate plot points
+smmry <- function(abc){
+    avg. <- mean(abc)
+    s.e. <- sd(abc) / sqrt(length(abc))
+    ans <- c(low=avg.-s.e., avg=avg., upp=avg.+s.e., se=s.e.)
+    return(ans)
+}
+estNames <- colnames(omegaEst)
+imposed <- apply(dndsEst, 1, smmry); dndsMat <- imposed[c(1,2,3),estNames]
+inferred <- apply(omegaEst, 2, smmry); omegaMat <- inferred[1:3,estNames]
+
+# ><>< # Estimate correlation coefficient of average estimates
+svCor <- cor(dndsMat["avg",],omegaMat["avg",])
+message( paste0("\nCorrelation Coefficient: ", sprintf("%.4f",svCor), "\n"))
+
+# ><>< # Summarise standard errors
+serror1 <- dndsMat["upp",] - dndsMat["avg",]
+serror2 <- omegaMat["upp",] - omegaMat["avg",]
+seTxt <- function(x.) paste(sprintf("%.4f",range(x.)), collapse=", ")
+message( paste0("\nS.E. Range (dN/dS): ", seTxt(serror1), "\n"))
+message( paste0("\nS.E. Range (omega): ", seTxt(serror2), "\n"))
+
+# ><>< # Process Position
+pdf(imgfile, height=15, width=15) 
+  par(mar=c(0, 0, 0, 0))
+  plot(0, 0, col="white", bty="n", xlab="", ylab="",
+    xlim=c(.1,20.2), ylim=c(.32,20.8), xaxt="n", yaxt="n")
+
+text(10.5, -0.15, expression(paste("OU Mean Reversion Rate, ",theta)), cex=2.5)
+text(-.4, 10.5, expression(paste("dN/dS and ",omega)), cex=2.5, srt=90)
+zWords <- expression(paste("OU Asymptotic Variance, ", Sigma^{2}))
+text(10.5, 21.2, zWords, cex=2.5)
+
+abX <- function(ygap) for(i in seq(1,20)) lines(rep(i,2), ygap, col="azure2")
+abX(c(.7, 5.2)); abX(c(5.8, 10.2)); abX(c(10.8, 15.2)); abX(c(15.8, 20.15))
+
+abY <- function(xgap) for(i in seq(1,20)) lines(xgap, rep(i,2), col="azure2")
+abY(c(.8, 5.2)); abY(c(5.8, 10.2)); abY(c(10.8, 15.2)); abY(c(15.8, 20.2))
+
+hLyne <- function(x1, x2, y0) lines(c(x1,x2), rep(y0,2), lwd=2)
+phold <- sapply(seq(.8,16,by=5), function(a) hLyne(a, a+4.4, .8))
[…]
