### 2026-02-02 #9797 @thsadiq
@clauswilke thanks again for your comments. See my responses below.

- Differentiation of analytical and likelihood inferences of natural selection is now stated explicitly. See for example, Line 52, Line 54 and Figure 3 caption.

- Lines 55 to 58 have been slightly edited to minimise misunderstanding.

- A sentence has been added to Line 78 stating the path to a more comprehensive supplementary folder. The associated readme file has also been edited to clearly imply what data correspond to which image.

- Analytical estimates of selection effect are returned as part of the package outputs. This is illustrated in Figure 1. It is also explained in Page 14 of the [package reference manual](https://bioconductor.org/packages/release/bioc/manuals/scoup/man/scoup.pdf).

Please let me know if these answer your queries or if there are any other reservations you have. I really appreciate your time and effort.
