### 2026-02-02 #9797 @jromanowska
Thank you, @clauswilke and @thsadiq, for the work. @thsadiq - the reviewer is correct, please, merge the joss branch into main (or copy the changes).

@ogoeli, could you please confirm if you have completed the review and are satisfied with the submission?
