### 2026-02-03 567c063
Update DESCRIPTION
JOSS publication update
