+++ DESCRIPTION
@@ -1,11 +1,16 @@
 Package: scoup
 <<<<<<< HEAD
+<<<<<<< HEAD
 Version: 1.5.0
 Date: 2025-01-10
 =======
 Version: 1.5.1
 Date: 2026-02-03
 >>>>>>> 3111287 (Update DESCRIPTION)
+=======
+Version: 1.5.1
+Date: 2026-02-03
+>>>>>>> eedb460 (Update DESCRIPTION)
 License: GPL (>= 2)
 Authors@R: c(person("Hassan", "Sadiq", role=c("aut","cre","cph"),
                     email="hassan.t.sadiq@gmail.com",
