+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: scoup
-Version: 1.1.4
-Date: 2025-01-10
+Version: 1.5.1
+Date: 2026-02-03
 License: GPL (>= 2)
 Authors@R: c(person("Hassan", "Sadiq", role=c("aut","cre","cph"),
                     email="hassan.t.sadiq@gmail.com",
