### 2026-02-03 040efd1
Update paper.md
JOSS publication update
