### 2026-02-03 b3b0e39
Reconcile README.md
