### 2026-02-03 b42c9ca
Reconcile paper.md
