### 2026-02-03 60d8fb5
Reconcile vignettes folder
