+++ paper/paper.md
@@ -109,6 +109,16 @@ averaged over all selection coefficient updates at each site and across
 the alignment. Likelihood inferences of $\omega$ were obtained with `CODEML`
 in `PAML` [@yang2007].
 
+![\label{pseudocode}**An example R code for simulating a codon sequence
+  alignment with `scoup`**. Default values were left unchanged. `Line01`:
+  OU adaptation parameters where, $\mu=0$, $\Sigma^{2}_{}=0.01$ and
+  $\theta=0.01$. `Line02`: evolution model input where,
+  $\mathbf{s} \sim \text{Gamma}(1,\sigma_{n}^{-1})$,
+  $\sigma^{2}_{n}=10^{-5}_{}$, $\sigma^{2}_{s}=10^{-5}_{}$ and
+  effective population size, $N_{\texttt{e}}^{}=1000$. `Line03`:
+  sequence information where, site count is $250$, extant taxa count
+  is $64$ and branch length is $0.1$.](FIG1.pdf)
+  
 ![\label{testfig}**Demonstration of the accuracy of outputs from `scoup`
   in terms of the likelihood $\omega$ and the analytical dN/dS measures
   of natural selection.** The estimates of the selection measures were
