### 2026-02-03 #9797 @thsadiq
Thank you @clauswilke and @ogoeli. I am heartily grateful for the times spent reviewing the paper and for the positive feedback. @jromanowska thank you for ensuring that the review process get finalised promptly. I have merged the joss and main git branches as advised. I have also completed the following post-review tasks.

- Double checked that authors and affiliations (including ORCIDs) are correct.

- The `scoup` package has been released on Bioconductor. The most recent post-review version number is **1.5.1** and it is accessible on the development node of the (host platform)[https://bioconductor.org/packages/devel/bioc/html/scoup.html].

- The release version 1.5.1 is archived on Zenodo and I retained the following DOI from Bioconductor: 10.18129/B9.bioc.scoup.

- The author list, ORCID and the licence from the Zenodo archive match those in the JOSS paper.

Please let me know if there is any other thing required from me.
