### 2026-02-05 #9797 @jromanowska
@thsadiq, I have just read your paper and the bioRxiv paper that you cite and noticed that they include almost the same text.  I think the paper should be completely re-written, if any self-plagiarism is to be mitigated.

@Kevin-Mattheus-Moerman - could you take a look at this?
