### 2026-02-05 #9797 @thsadiq
@jromanowska thank you for taking the time to thoroughly read through the paper and the preprint. Indeed the JOSS manuscript is a summarized version of the bioRxiv preprint. I expect that, after publication, the bioRxiv version will update to reflect the JOSS article as the correct citation. The preprint was used as an avenue to expand on some of the concepts. I am not sure that a rewrite is necessary given that the titles and authors are the same.
