### 2026-02-05 #9797 @thsadiq
@jromanowska here is a an example of a JOSS paper that has a preprint on arxiv. arxiv: https://arxiv.org/abs/2406.07817 and JOSS: https://joss.theoj.org/papers/10.21105/joss.06889
