### 2026-02-05 #9797 @clauswilke
@thsadiq If you intend the arxiv version to be a preprint for the JOSS paper then maybe update the arxiv version to the latest manuscript version submitted to JOSS? In this case there is no doubt about your intent.
