### 2026-02-05 #9797 @jromanowska
I assumed the paper on bioRxiv was a different one, since you cite it in JOSS paper. I will still have some comments for the JOSS paper, since I think the text doesn't read well. But now it's late, so I'll get back to you one of the next days :)
