### 2026-02-06 #9797 @thsadiq
Thank you @jromanowska. I will wait for your feedback on the JOSS paper. If necessary, I am happy to update the biorxiv version so that it is the same as the JOSS manuscript as recommended by @clauswilke. The biorxiv version was originally intended as an expanded version of the JOSS paper and that is why it was only mentioned as a whitepaper and not listed as a reference. I now realise how that may be inappropriate.
