### 2026-02-09 #9797 @jromanowska
@thsadiq, thank you for your patience. I think that you need to consider actually creating two papers: one that is focused on the maths and results of the runs (this might be on arXiv) and one that is focused on coding and implementation details (for JOSS). Currently, your JOSS submission is difficult to understand because it lacks some explanations and it focuses on results of the runs, not on how you actually created the package. The "Implementation" section is actually "Results" or "Exemplary run" section. I'd like to see more explanation of the design choices you needed to make when writing the package.

Thus, I suggest that you keep the arXiv paper as is (and may cite it in the JOSS paper), but re-write JOSS paper to include only "Summary", "Statement of need", the actual details of implementation (and perhaps "Conclusions"). However, please try not to copy entire chunks from the arXiv paper.

I hope you can understand this decision.
