### 2026-02-09 #9797 @thsadiq
@jromanowska, thanks again for your feedback. I will begin work on your comments. I have one question that I would like if you could kindly clarify. Is it fine to retain the same titles for both the JOSS and the bioRxiv papers as it is currently the case?
