### 2026-02-17 1a99736
Incorporating comment from JOSS editor
