+++ paper/tex/FIG0.tex
@@ -1,138 +1,0 @@
-% ><>< ================================================================= ><>< %
-% ><><     scoup:- Simulate Codon Sequences with Darwinian Selection     ><>< %
-% ><><           Incorporated as an Ornstein-Uhlenbeck Process           ><>< %
-% ><><                       ~~~~~~~~~~~~~~~~~~~~~                       ><>< %
-% ><><                          13th June 2025.                          ><>< %
-% ><>< ================================================================= ><>< %
-
-\documentclass[12pt]{article}
-
-% ><><                      Load necessary packages                      ><>< %
-\usepackage{tikz}
-\usepackage{pifont}
-\usepackage{multirow}
-\usepackage{enumerate}
-\usepackage{graphicx, color}
-\usepackage[hidelinks]{hyperref}
-\usepackage{amsmath, amsthm, amsfonts, amssymb}
-\DeclareMathAlphabet{\mathpzc}{OT1}{pzc}{m}{it}
-\usepackage[paperheight=4.28in, paperwidth=10.4in,%10.9in,
-    margin=.1in, heightrounded]{geometry}
- \renewcommand{\familydefault}{\sfdefault}
-% --------------------------------------------------------------------------- %
-
-% ><><                          Define colours                           ><>< %
-\definecolor{d00}{rgb}{0.00, 0.00, 0.00}
-\definecolor{d02}{rgb}{0.80, 0.80, 0.80}
-\definecolor{d03}{rgb}{0.13, 0.55, 0.13}
-\definecolor{d04}{rgb}{0.50, 0.50, 0.50}
-\definecolor{d05}{rgb}{1.00, 1.00, 1.00}
-% --------------------------------------------------------------------------- %
-
-% ><><                      Tikz shapes and designs                      ><>< %
-\usetikzlibrary{positioning}
-\usetikzlibrary{shapes, arrows}
-\usetikzlibrary{shapes.gates.logic.US}
-\usetikzlibrary{decorations.pathreplacing}
-\usetikzlibrary{decorations.pathmorphing,patterns}
-\usetikzlibrary{calc, patterns, decorations.markings}
-
-% ><>< ==== ><>< % Lines
-\tikzstyle{Line} = [draw, line width=.12cm, rounded corners, -latex']
-% ><>< ==== ><>< % Rectangles
-\tikzstyle{arect} = [draw, rectangle, minimum width=-0.7em,
-    minimum height=2em, text centered, rounded corners,
-    fill=white, line width=.08cm]
-% ><>< ==== ><>< % Gate
-\tikzstyle{on}=[draw, chamfered rectangle,
-    white, fill=d03, double=d03, very thick,
-    text width=4em, minimum height=3em, text centered, line width=.08cm]
-\tikzstyle{off}=[draw, chamfered rectangle,
-    white, fill=red, double=red, very thick,
-    text width=4em, minimum height=3em, text centered, line width=.08cm]
-\tikzstyle{elpz} = [draw, ellipse, line width=.1cm]
-% ><>< ==== ><>< % Tape
-\tikzstyle{dcmt} = [draw, tape, minimum width=6cm,
-    d00, tape bend height=0.4cm, minimum height=2em,
-    text centered, fill=white, tape bend top = none, line width=.08cm]
-% ><>< ==== ><>< % Trapezoid
-\tikzstyle{traped} = [draw, trapezium, minimum width=-5em, d00,
-    trapezium left angle=97, trapezium right angle=83,
-    minimum height=2em, text centered, fill=white, line width=.08cm]
-% ><>< ==== ><>< % Cylinder
-\tikzstyle{clnd} = [draw, cylinder, minimum width=-2em,
[…]
