### 2026-02-17 d2cdab7
Incorporating comment from JOSS editor
