### 2026-02-17 #9797 @thsadiq
@jromanowska thanks again for your comments. We opted to adopt the suggestion by @clauswilke. Therefore, we replaced the bioRxiv preprint so that it is identical to the submitted JOSS manuscript. This ensures that we do not have to worry about title change. It also avails us enough time to complete our ongoing projects that rely on the package where we intend to incorporate and elaborate on the results in our initial submission. In response to the requested clarifications (a.) the result section is removed and (b.) the algorithm and implementation sections are updated. Kindly let me know if you have any further comments while we await approval of the updated bioRxiv preprint.
