### 2026-02-18 367da9c
Some fine-tuning in-line with JOSS updates
