### 2026-02-18 0aa34a0
Some fine-tuning in-line with JOSS review
