+++ README.md
@@ -12,7 +12,7 @@ A R package intended for the Bioconductor platform. It avails an opportunity to 
 - [man](man): a folder where the R markdown files are saved.
 - [NAMESPACE](NAMESPACE): what functions and/or classes are imported and/or exported.
 - [NEWS](NEWS): Details of the edits, bug fixes and updates made to the package. It is consistent with the most recent version iteration on Bioconductor.
-- [paper](paper): contains the files used to generate the manuscript about the package to the Journal of Open Source Software. Similar files for other publications related to the package will be included herein as they become available.
+- [paper](paper): contains the files that are used to prepare the manuscript that describes the package for the Journal of Open Source Software. Similar files for other publications related to the package will be included herein as they become available.
 - [R](R): a folder of R scripts that contains different (back-end) functions of the package.
 - [README](README.md): the file that contains this preface information.
 - [tests](tests): a directory of R scripts where sample applications of the package are saved. The scripts are intended to assist users get familiar with the package. They are also (minimally) used within the vignette.
