### 2026-02-18 #9797 @jromanowska
I have some comments to the manuscript text:

- lines 24-25: I don't understand this sentence, could you re-write it? And please, explain, how does the citation (Jacques et al., 2023) fit to this sentence - I've scanned the paper and it seems only partially relevant.
- line 26: You write "Validation of these largely codon-based models" - which models? You don't mention any models in the previous sentence.
- next sentence: you mention two references, so two models to back up the statement "given the ever increasing diversity of natural selection inference models that exist" - is there possibly an article with an overview of models? This would suit much better.
- lines 30-31: Please, correct: "Bioconductor is a leading platform where peer-reviewed bioinformatic software useful for biological data analyses are hosted" -> "Bioconductor is a leading bioinformatics platform distributing peer-reviewed R packages."
- lines 32-33: maybe you want to update the search results here?
- lines 36-37: This is misleading: after "Software and/or packages for simulating genetic sequences are also rare in the scientific literature", you cite a CRAN Task view, which focuses on R packages, not 'scientific literature'. (and the newest version now is from 2025)
- Caption to Fig.1: From the schematic, it seems like there is a linear path from START to STOP, while the caption suggests that there might be a cycle ("After each substitution event, the process returns to STEP A") - please, explain.
- line 51: what is "seascape setting"? (_I'm not an expert in evolution algorithms, sorry_)
- your section named "Implementation" is actually a use case, not implementation - please either remove that (as I can see that you have much better example in the vignette on Bioconductor) or write about implementation: what kind of design you chose when writing the code, any performance metrics, or any special dependencies or requirements. _(Also: when writing a code, it's much better to include it as code, not as figures)_
- line 60-62: Rewrite to e.g.,: "We present `scoup`, an R package for codon sequences simulation, where the evolutionary processes are mirrored more realistically than most existing simulators" _(and, importantly: you haven't mentioned any other simulators: please, change this or add some references above and here)_
- lineas 65-67: What does this sentence mean: "In another unique contribution to the literature, the magnitude of the Darwinian selection affect on the simulated sequences was controlled with the ratio of the variances of selection coefficients."?
- After correcting the points above, please ask a colleague, who is not directly involved in the project, to read the entire paper for you.
