### 2026-02-18 b3a6b1a
Edited paper to accomodate editor's comments
