+++ paper/paper.md
@@ -53,24 +53,39 @@ populations.
 
 # Statement of need
 
-Statistical inference of the extent to which Darwinian natural selection has
-impacted genetic data, commands a healthy portion of the phylogenetic
-literature [@jacques2023]. Validation of these largely codon-based models
-relies heavily on simulated data. Given the ever increasing diversity of
-natural selection inference models that exist [@yang2007; @hyphy2020],
-there is a need for more sophisticated simulators to match the
-expanding model complexities.
+Statistical inference models that are used to analyse the degree of the impact
+of Darwinian natural selection on observed genetic data, command a healthy
+portion of the phylogenetic literature [@gupta2023]. Validation of these
+largely codon-based models relies heavily on simulated data. Given the ever
+increasing diversity of natural selection inference models that exist
+[@yang2007; @arenas2015; @hyphy2020], there is a need for more sophisticated
+simulators to match the expanding model complexities.
 
-Bioconductor [@bioc2004] is a leading platform where peer-reviewed
-bioinformatic software useful for biological data analyses are hosted. A
-search of the entries on the platform, in Version 3.19 on 29 October 2024,
-with keywords including, `codon`, `mutation`, `selection`, `simulate`, and
-`simulation` returned a total of 72 unique packages out of the 2300 available.
-None of the retrieved entries was dedicated to codon data simulation for
-natural selection analyses. Thus, `scoup` is designed on the basis of the
-mutation-selection (MutSel) framework [@halpern1998] as an overdue
-contribution to the void. Software and/or packages for simulating genetic
-sequences are also rare in the scientific literature [@gearty2024]. 
+Bioconductor [@bioc2004] is a leading bioinformatics platform distributing
+peer-reviewed R packages. A search of the entries on the platform, in
+Version 3.22 on 18 February 2026, with keywords including, `codon`, `mutation`,
+`selection`, `simulate`, and `simulation` returned a total of 70 packages
+(excluding `scoup`) out of the 2361 available. None of the retrieved entries
+was dedicated to codon data simulation for natural selection analyses. Thus,
+`scoup` is designed on the basis of the mutation-selection (MutSel) framework
+[@halpern1998] as an overdue contribution to the void.
+
+Software and/or packages for simulating molecular protein sequences are
+a few in the scientific literature [@peng2015]. Existing simulators tend
+to be more suitable for quantitative character evolution. These include,
+`ape` [@paradis2019], `ouch` [@cressler2015; @butler2004] and `geiger`
+[@pennell2014]. Other extensively used DNA sequence simulators including,
+`Seq-Gen` [@rambaut1997], `INDELible` [@fletcher2009], `PhyloSim`
+[@sipos2011] and `phangorn` [@klaus2011] are parameterized in accordance
+with $\omega$-based models [@goldman1994; @muse1994]. More recent sequence
+simulators, such as, `phastSim` [@nicola2022] and `AliSim-HPC` [@nhan2023]
+prioritized output capacity. Only few genetic simulators were built upon the
+more elaborate MutSel evolutionary concept. These include, `Pyvolve`
+[@wilke2015] and `SGWE` [@arenas2014]. To the best of our knowledge, these
+existing MutSel friendly simulators are only able to generate data from
+static landscapes. With our proposed simulator (`scoup`), it is possible
+to generate codon sequences from landscapes that are static or those that
+are changing (also known as *seascapes*) [@lassig2009].
 
 # Algorithm
 
@@ -85,8 +100,9 @@ biologically amenable, Ornstein-Uhlenbeck (OU) process
[…]
