+++ paper/paper.bib
@@ -1,23 +1,16 @@
-% ><>< ========================================================================= ><>< %
-% ><><   scoup: Simulate Codon Sequences with Darwinian Selection Incorporated   ><>< %
-% ><><                     as an Ornstein-Uhlenbeck Process.                     ><>< %
-% ><><                           ~~~~~~~~~~~~~~~~~~~~~                           ><>< %
-% ><><                               Bibliography.                               ><>< %
-% ><><                           ~~~~~~~~~~~~~~~~~~~~~                           ><>< %
-% ><><                              13th June 2025.                              ><>< %
-% ><>< ========================================================================= ><>< %
+% ><>< ================================================================ ><>< %
+% ><><     scoup: Simulate Codon Sequences with Darwinian Selection     ><>< %
+% ><><          Incorporated as an Ornstein-Uhlenbeck Process.          ><>< %
+% ><>< ================================================================ ><>< %
 
-@article{jacques2023,
-  title   = "{Roadmap to the Study of Gene and Protein Phylogeny and Evolution
-            -- A Practical Guide}",
-  author  = {Florian Jacques and Paulina Bolivar and Kristian Pietras and
-            Emma U Hammarlund},
-  journal = "{PLoS One}",
-  year    = {2023},
-  pages   = {e0279597},
-  volume  = {18},
-  number  = {2},
-  doi     = {10.1371/journal.pone.0279597}
+@article{gupta2023,
+  title =   "{Next-Generation Development and Application of Codon Model in
+             Evolution}",
+  author =  {Manoj Kumar Gupta and Ramakrishna Vadde},
+  journal = {Frontiers in Genetics},
+  year =    {2023},
+  volume =  {14},
+  doi =     {10.3389/fgene.2023.1091575},
 }
 
 @article{yang2007,
@@ -32,13 +25,13 @@
 }
 
 @article{hyphy2020,
-  title   = "{{HyPhy 2.5} -- A Customizable Platform for Evolutionary Hypothesis
-              Testing Using Phylogenies}",
-  author  = {Sergei L {Kosakovsky Pond} and Art F Y Poon and Ryan Velazquez and
-             Steven Weaver and N Lance Hepler and Ben Murrell and Stephen D Shank
-             and Brittany Rife Magalis and Dave Bouvier and Anton Nekrutenko and
-             Sadie Wisotsky and Stephanie J Spielman and Simon D W Frost
-             and S V Muse},
+  title   = "{{HyPhy 2.5} -- A Customizable Platform for Evolutionary
+              Hypothesis Testing Using Phylogenies}",
+  author  = {Sergei L {Kosakovsky Pond} and Art F Y Poon and Ryan Velazquez
+             and Steven Weaver and N Lance Hepler and Ben Murrell and Stephen
+             D Shank and Brittany Rife Magalis and Dave Bouvier and Anton
+             Nekrutenko and Sadie Wisotsky and Stephanie J Spielman and
+             Simon D W Frost and S V Muse},
   journal = {Molecular Biology and Evolution},
   year    = {2020},
   pages   = {295--299},
@@ -51,13 +44,13 @@
   title  = "{Bioconductor: Open Software Development for Computational Biology
              and Bioinformatics}",
   author  = {Robert C Gentleman and Vincent J Carey and Douglas M Bates and
-             Ben Bolstad and Marcel Dettling and Sandrine Dudoit and Byron Ellis
[…]
