+++ paper/tex/FIG3.tex
@@ -1,56 +1,0 @@
-% ><>< ================================================================= ><>< %
-% ><><     scoup:- Simulate Codon Sequences with Darwinian Selection     ><>< %
-% ><><           Incorporated as an Ornstein-Uhlenbeck Process           ><>< %
-% ><><                       ~~~~~~~~~~~~~~~~~~~~~                       ><>< %
-% ><><                          13th June 2025.                          ><>< %
-% ><>< ================================================================= ><>< %
-
-\documentclass[12pt]{article}
-
-% ><><                      Load necessary packages                      ><>< %
-\usepackage{pifont}
-\usepackage{multirow}
-\usepackage{enumerate}
-\usepackage{graphicx, color}
-\usepackage[hidelinks]{hyperref}
-\usepackage{amsmath, amsthm, amsfonts, amssymb}
-\DeclareMathAlphabet{\mathpzc}{OT1}{pzc}{m}{it}
-\usepackage[paperheight=1.9in, paperwidth=10.05in,
-    margin=.1in, heightrounded]{geometry}
- \renewcommand{\familydefault}{\sfdefault}
-% --------------------------------------------------------------------------- %
-
-% ><><                         Begin compilation                         ><>< %
-\begin{document}
-\thispagestyle{empty}
-
-\newcommand{\xbig}{\fontsize{40}{0}\selectfont}
-\setlength{\fboxrule}{2.5pt}
-\setlength{\fboxsep}{0pt}
-
-\begin{minipage}{\hsize}\xbig
-
-\fbox{
-   \centering
-   {\renewcommand{\tabcolsep}{0.30cm}
-   \renewcommand{\arraystretch}{1.1}
-   \hspace*{-.9cm}
-   \begin{tabular}{l}
-        \texttt{adaptEnrty} \texttt{<-} \texttt{ouInput()}
-            \texttt{\#} \texttt{Line01} \rule{0ex}{2.2ex}\\[1ex]
-        \texttt{modelEntry} \texttt{<-} \texttt{hbInput()}
-            \texttt{\#} \texttt{Line02} \\[1ex]
-        \texttt{sqEntry} \texttt{<-} \texttt{seqDetails()}
-            \texttt{\#} \texttt{Line03} \\[1ex]
-        \texttt{seqData} \texttt{<-} \texttt{alignsim(adaptEnrty,}
-            \texttt{sqEntry,} \texttt{modelEntry)} \\[1ex]
-   \end{tabular}}
-}
-\end{minipage}
-
-
-\end{document}
-
-% ><>< ================================================================= ><>< %
-% ><><                        DOCUMENT ENDS HERE.                        ><>< %
-% ><>< ================================================================= ><>< %

[…]
