### 2026-02-18 0fb4321
Edited paper to accomodate editor's comments
