+++ paper/paper.bib
@@ -4,13 +4,13 @@
 % ><>< ================================================================ ><>< %
 
 @article{gupta2023,
-  title =   "{Next-Generation Development and Application of Codon Model in
+  title   = "{Next-Generation Development and Application of Codon Model in
              Evolution}",
-  author =  {Manoj Kumar Gupta and Ramakrishna Vadde},
+  author  = {Manoj Kumar Gupta and Ramakrishna Vadde},
   journal = {Frontiers in Genetics},
-  year =    {2023},
-  volume =  {14},
-  doi =     {10.3389/fgene.2023.1091575},
+  year    = {2023},
+  volume  = {14},
+  doi     = {10.3389/fgene.2023.1091575},
 }
 
 @article{yang2007,
@@ -22,6 +22,17 @@
   number  = {8},
   pages   = {1586--1591},
   doi     = {10.1093/molbev/msm088}
+}
+
+@article{nhan2023,
+  title   = "{{AliSim-HPC}: Parallel Sequence Simulator for Phylogenetics}",
+  author  = {Nhan {Ly-Trong} and Giuseppe M J Barca and Bui Quang Minh},
+  journal = {Bioinformatics},
+  year    = {2023},
+  volume  = {39},
+  number  = {9},
+  pages   = {btad540},
+  doi     = {10.1093/bioinformatics/btad540},
 }
 
 @article{hyphy2020,
@@ -85,184 +96,185 @@
 }
 
 @article{arenas2015,
-  title =   "{Trends in Substitution Models of Molecular Evolution}",
-  author =  {Miguel Arenas},
+  title   = "{Trends in Substitution Models of Molecular Evolution}",
+  author  = {Miguel Arenas},
   journal = {Frontiers in Genetics},
-  year =    {2015},
-  volume =  {6},
-  number =  {319},
-  doi =     {10.3389/fgene.2015.00319},
+  year    = {2015},
+  volume  = {6},
+  number  = {319},
+  doi     = {10.3389/fgene.2015.00319},
 }
 
 @article{paradis2019,
-  title =   "{{ape 5.0}: An Environment for Modern Phylogenetics and
+  title   = "{{ape 5.0}: An Environment for Modern Phylogenetics and
               Evolutionary Analyses in R}",
-  author =  {Emmanuel Paradis and Klaus Schliep},
+  author  = {Emmanuel Paradis and Klaus Schliep},
   journal = {Bioinformatics},
-  year =    {2019},
-  volume =  {35},
-  number =  {3},
-  pages =   {526--528},
-  doi =     {10.1093/bioinformatics/bty633},
+  year    = {2019},
+  volume  = {35},
+  number  = {3},
+  pages   = {526--528},
+  doi     = {10.1093/bioinformatics/bty633},
 }
 
 @article{cressler2015,
-  title =   "{Detecting Adaptive Evolution in Phylogenetic Comparative
+  title   = "{Detecting Adaptive Evolution in Phylogenetic Comparative
               Analysis Using the Ornstein-Uhlenbeck Model}",
-  author =  {Clayton E Cressler and Marguerite A Butler and Aaron A King},
+  author  = {Clayton E Cressler and Marguerite A Butler and Aaron A King},
   journal = {Systematic Biology},
-  year =    {2015},
-  volume =  {64},
-  number =  {6},
-  pages =   {953--968},
-  doi =     {10.1093/sysbio/syv043},
+  year    = {2015},
+  volume  = {64},
+  number  = {6},
+  pages   = {953--968},
+  doi     = {10.1093/sysbio/syv043},
 }
 
 @article{butler2004,
-  title =   "{Phylogenetic Comparative Analysis: A Modeling Approach 
[…]
