### 2026-02-19 ef1d36b
Some fine-tuning in-line with JOSS review
