+++ paper/paper.md
@@ -38,18 +38,18 @@ evolutionary biology (i.e. phylogenetics and population genetics) borne
 from the split can only benefit from research that seeks to bridge
 the gap. Simulation algorithms that combine fundamental concepts from
 both genres are important to achieve such unifying objective. 
-We introduce `scoup`, a codon sequence simulator that is implemented
-in R and hosted on the Bioconductor platform. There is hardly any other
-simulator dedicated to genetic sequence generation for natural selection
-analyses on the platform. Concepts from the Halpern-Bruno mutation-selection
-model and the Ornstein-Uhlenbeck (OU) evolutionary algorithm were creatively
-fused such that the end-product is a novelty with respect to computational
-genetic simulation. Users are able to seamlessly adjust the model parameters
-to mimic complex evolutionary procedures that may have been otherwise
-infeasible. For example, it is possible to explicitly interrogate the
-concepts of static and changing fitness landscapes with regards to Darwinian
-natural selection in the context of codon sequences from multiple
-populations.
+We introduce [`scoup`](doi.org/10.18129/B9.bioc.scoup), a codon sequence
+simulator that is implemented in R and hosted on the Bioconductor platform.
+There is hardly any other simulator dedicated to genetic sequence generation
+for natural selection analyses on the platform. Concepts from the
+Halpern-Bruno mutation-selection model and the Ornstein-Uhlenbeck (OU)
+evolutionary algorithm were creatively fused such that the end-product is a
+novelty with respect to computational genetic simulation. Users are able to
+seamlessly adjust the model parameters to mimic complex evolutionary
+procedures that may have been otherwise infeasible. For example, it is
+possible to explicitly interrogate the concepts of static and changing
+fitness landscapes with regards to Darwinian natural selection in the
+context of codon sequences from multiple populations.
 
 # Statement of need
 
@@ -65,10 +65,11 @@ Bioconductor [@bioc2004] is a leading bioinformatics platform distributing
 peer-reviewed R packages. A search of the entries on the platform, in
 Version 3.22 on 18 February 2026, with keywords including, `codon`, `mutation`,
 `selection`, `simulate`, and `simulation` returned a total of 70 packages
-(excluding `scoup`) out of the 2361 available. None of the retrieved entries
-was dedicated to codon data simulation for natural selection analyses. Thus,
-`scoup` is designed on the basis of the mutation-selection (MutSel) framework
-[@halpern1998] as an overdue contribution to the void.
+(excluding [`scoup`](doi.org/10.18129/B9.bioc.scoup)) out of the 2361 available.
+None of the retrieved entries was dedicated to codon data simulation for natural
+selection analyses. Thus, [`scoup`](doi.org/10.18129/B9.bioc.scoup) is designed
+on the basis of the mutation-selection (MutSel) framework [@halpern1998] as an
+overdue contribution to the void.
 
 Software and/or packages for simulating molecular protein sequences are
 a few in the scientific literature [@peng2015]. Existing simulators tend
@@ -83,21 +84,22 @@ prioritized output capacity. Only few genetic simulators were built upon the
 more elaborate MutSel evolutionary concept. These include, `Pyvolve`
 [@wilke2015] and `SGWE` [@arenas2014]. To the best of our knowledge, these
 existing MutSel friendly simulators are only able to generate data from
-static landscapes. With our proposed simulator (`scoup`), it is possible
-to generate codon sequences from landscapes that are static or those that
-are changing (also known as *seascapes*) [@lassig2009].
+static landscapes. With our proposed simulator, it is possible to generate
[…]
