+++ paper/paper.bib
@@ -24,15 +24,15 @@
   doi     = {10.1093/molbev/msm088}
 }
 
-@article{nhan2023,
-  title   = "{{AliSim-HPC}: Parallel Sequence Simulator for Phylogenetics}",
-  author  = {Nhan {Ly-Trong} and Giuseppe M J Barca and Bui Quang Minh},
-  journal = {Bioinformatics},
-  year    = {2023},
-  volume  = {39},
-  number  = {9},
-  pages   = {btad540},
-  doi     = {10.1093/bioinformatics/btad540},
+
+@article{arenas2015,
+  title   = "{Trends in Substitution Models of Molecular Evolution}",
+  author  = {Miguel Arenas},
+  journal = {Frontiers in Genetics},
+  year    = {2015},
+  volume  = {6},
+  number  = {319},
+  doi     = {10.3389/fgene.2015.00319},
 }
 
 @article{hyphy2020,
@@ -93,16 +93,6 @@
   number  = {1},
   pages   = {2--10},
   doi     = {10.1002/gepi.21876}
-}
-
-@article{arenas2015,
-  title   = "{Trends in Substitution Models of Molecular Evolution}",
-  author  = {Miguel Arenas},
-  journal = {Frontiers in Genetics},
-  year    = {2015},
-  volume  = {6},
-  number  = {319},
-  doi     = {10.3389/fgene.2015.00319},
 }
 
 @article{paradis2019,
@@ -241,6 +231,17 @@
   doi     = {10.1371/journal.pcbi.1010056},
 }
 
+@article{nhan2023,
+  title   = "{{AliSim-HPC}: Parallel Sequence Simulator for Phylogenetics}",
+  author  = {Nhan {Ly-Trong} and Giuseppe M J Barca and Bui Quang Minh},
+  journal = {Bioinformatics},
+  year    = {2023},
+  volume  = {39},
+  number  = {9},
+  pages   = {btad540},
+  doi     = {10.1093/bioinformatics/btad540},
+}
+
 @article{wilke2015,
   title   = "{Pyvolve: A Flexible Python Module for Simulating Sequences
              along Phylogenies}",
@@ -323,3 +324,19 @@
   doi     = {10.1093/genetics/148.3.929}
 }
 
+@manual{bates2024,
+  title   = "{Matrix: Sparse and Dense Matrix Classes and Methods}",
+  author  = {Douglas Bates and Martin Maechler and Mikael Jagan},
+  year    = {2024},
+  note    = {R package version 1.7-0},
+  doi     = {10.32614/CRAN.package.Matrix}
+}
+
+@manual{pages2024,
+  title   = "{Biostrings: Efficient Manipulation of Biological Strings}",
+  author  = {Herv\'e Pag\`es and Patrick Aboyoun and Robert Gentleman and
+             Saikat {DebRoy}},
+  year    = {2024},
+  note    = {R package version 2.73.1},
+  doi     = {10.18129/B9.bioc.Biostrings}
+}
