### 2026-02-19 ba5ad3c
Some fine-tuning in-line with JOSS review
