+++ paper/paper.bib
@@ -312,18 +312,6 @@
   doi     = {10.1103/PhysRev.36.823}
 }
 
-@article{nielsen1998,
-  title   = "{Likelihood Models for Detecting Positively Selected Amino
-            Acid Sites and Applications to the HIV-1 Envelope Gene}",
-  author  = {Rasmus Nielsen and Ziheng Yang},
-  journal = {Genetics},
-  year    = {1998},
-  volume  = {148},
-  number  = {3},
-  pages   = {929--936},
-  doi     = {10.1093/genetics/148.3.929}
-}
-
 @manual{bates2024,
   title   = "{Matrix: Sparse and Dense Matrix Classes and Methods}",
   author  = {Douglas Bates and Martin Maechler and Mikael Jagan},
