### 2026-02-19 365b08a
Some fine-tuning in-line with JOSS review
