+++ .github/workflows/draft.yml
@@ -1,30 +1,0 @@
-name: Draft PDF
-on:
-  push:
-    paths:
-      - paper/paper.md
-      - paper/paper.bib
-      - paper/README.md
-      - .github/workflows/draft-pdf.yml
-      
-jobs:
-  paper:
-    runs-on: ubuntu-latest
-    name: Paper Draft
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-      - name: Build draft PDF
-        uses: openjournals/openjournals-draft-action@master
-        with:
-          journal: joss
-          # This should be the path to the paper within your repo.
-          paper-path: paper/paper.md
-      - name: Upload
-        uses: actions/upload-artifact@v4
-        with:
-          name: paper
-          # This is the output path where Pandoc will write the compiled
-          # PDF. Note, this should be the same directory as the input
-          # paper.md
-          path: paper/paper.pdf
