+++ paper/data/molecules/simtree.txt
@@ -1,0 +1,4 @@
+(((S001:0.1,S002:0.1):0.1,(S003:0.1,S004:0.1):0.1):0.1,((S005:0.1,S006:0.1):0.1,(S007:0.1,S008:0.1):0.1):0.1):0.1;
+
+
+

[…]
