### 2026-03-02 16166eb
Edited paper for JOSS submission
