+++ paper/paper.bib
@@ -278,6 +278,16 @@
   doi     = {10.1073/pnas.0907953107},
 }
 
+@article{wright1932,
+  title   = "{The Roles of Mutation, Inbreeding, Crossbreeding
+              and Selection in Evolution}",
+  author  = {Sewall Wright},
+  journal = {Proceedings of the Sixth International Congress of Genetics},
+  year    = {1932},
+  volume  = {1},
+  pages   = {356--366},
+}
+
 @article{spielman2015,
   title   = "{The Relationship between dN/dS and Scaled Selection
             Coefficients}",
