### 2026-03-03 #9797 @thsadiq
@jromanowska thanks for your insightful comments. Your suggestions have been incorporated to the newest version of the article. See below for point-by-point responses.

- [x] lines 24-25: I don't understand this sentence, could you re-write it? And please, explain, how does the citation (Jacques et al., 2023) fit to this sentence - I've scanned the paper and it seems only partially relevant.
  - Jacques et al. was not the intended citation. Sentence and citation have been updated.

- [x] line 26: You write "Validation of these largely codon-based models" - which models? You don't mention any models in the previous sentence.
  - The clarification incorporated in Lines 24 - 25 in response to the comment above also clarifies this. Replaced "Statistical inference ..." with "Statistical inference models ..."

- [x] next sentence: you mention two references, so two models to back up the statement "given the ever increasing diversity of natural selection inference models that exist" - is there possibly an article with an overview of models? This would suit much better.
  - Included a review article as advised. Retained the previous citations because they refer to two of the popular platforms that host commonly used natural selection inference models.

- [x] lines 30-31: Please, correct: "Bioconductor is a leading platform where peer-reviewed bioinformatic software useful for biological data analyses are hosted" -> "Bioconductor is a leading bioinformatics platform distributing peer-reviewed R packages."
  - Edited as advised.

- [x] lines 32-33: maybe you want to update the search results here?
  - Search updated. The lower updated total is most probably because packages that are not properly maintained on the platform are deprecated.

- [x] lines 36-37: This is misleading: after "Software and/or packages for simulating genetic sequences are also rare in the scientific literature", you cite a CRAN Task view, which focuses on R packages, not 'scientific literature'. (and the newest version now is from 2025)
  - Citation and text have been edited.

- [x] Caption to Fig.1: From the schematic, it seems like there is a linear path from START to STOP, while the caption suggests that there might be a cycle ("After each substitution event, the process returns to STEP A") - please, explain.
  - Edited the caption to indicate that the illustration is for a single substitution event.

- [x] line 51: what is "seascape setting"? (I'm not an expert in evolution algorithms, sorry)
  - It is when evolution is assumed to have occurred from changing fitness landscapes. This has been clarified with the last sentence of a paragraph added to **Statement of Need**.

- [x] your section named "Implementation" is actually a use case, not implementation - please either remove that (as I can see that you have much better example in the vignette on Bioconductor) or write about implementation: what kind of design you chose when writing the code, any performance metrics, or any special dependencies or requirements. (Also: when writing a code, it's much better to include it as code, not as figures)
  - Updated the section to highlight that the imported packages. However, it is important that some results are included in the manuscript because the [bioRxiv update](https://doi.org/10.1101/2025.06.14.659628) was rejected otherwise.

- [x] line 60-62: Rewrite to e.g.,: "We present scoup, an R package for codon sequences simulation, where the evolutionary processes are mirrored more realistically than most existing simulators" (and, importantly: you haven't mentioned any other simulators: please, change this or add some references above and here)
  - Edited accordingly and a review of existing molecular simulation packages has been summarised into a paragraph and added to the **Statement of Need** section.

- [x] lines 65-67: What does this sentence mean: "In another unique contribution to the literature, the magnitude of the Darwinian selection affect on the simulated sequences was controlled with the ratio of the variances of selection coefficients."?
  - Deleted sentence. Inserted it with more clarification as lines 89 - 93.

- [x] After correcting the points above, please ask a colleague, who is not directly involved in the project, to read the entire paper for you.
  - Updated version was reviewed and passed by two colleagues that are not experts in computational molecular evolution analyses or involved in the project.

Thanks again for taking the time to read through the manuscript thoroughly. I look forward to your feedback.
