### 2026-03-03 #9797 @jromanowska
After finishing with the paper, I'd like you to update some things online:

- [x] put the link to the bioconductor scoup-page on GitHub (under "About", on the right-hand side)
- [x] make sure the affiliations and author names are the same in the paper and on zenodo
- [x] update the released version on zenodo (I can see 1.5.1 there, while 1.5.2 on bioconductor)
