+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: scoup
-Version: 1.5.2
+Version: 1.5.3
 Date: 2026-02-03
 License: GPL (>= 2)
 Authors@R: c(person("Hassan", "Sadiq", role=c("aut","cre","cph"),
