### 2026-03-03 7ab288f
Final JOSS fine-tuning
