### 2026-03-03 #9797 @jromanowska
Thank you for incorporating changes to the manuscript.

I can't see that you made the above changes.

The author names and affiliations on zenodo must match those in the paper. Right now, it's:

| zenodo | paper |
|-------------|----------|
| Hassan Sadiq (1) | Hassan Sadiq (1,2) |
| Darren Martin | Darren P. Martin |

The versions must match. Right now, it's:

- 1.5.4 on [zenodo](https://zenodo.org/records/18469170)
- 1.4.0 on [bioconductor released version 3.22](https://bioconductor.org/packages/release/bioc/html/scoup.html)
- 1.5.2 on [bioconductor devel version 3.23](https://bioconductor.org/packages/devel/bioc/html/scoup.html)

Moreover, the button on scoup GitHub repo links to the released version, which is not the updated one.
