### 2026-03-03 45b843f
Final JOSS fine-tuning
