+++ README.md
@@ -2,12 +2,10 @@
 Simulate Codon Sequences with Darwinian Selection Incorporated as an Ornstein-Uhlenbeck Process
 
 ## About
-A R package intended for the Bioconductor platform. It avails an opportunity to simulate molecular data that are primarily useful for investigating the effects of Darwinian natural selection. Concepts from the population genomics and the phylogenetics literature are merged so that components of observed selection signatures in genetic sequences may be interrogated more thoroughly. This is achieved by combining (a.) the explicit use of selection coefficients by Halpern-Bruno in their popular 1998 mutation-selection codon model and (b.) the versatility of the stochastic Ornstein-Uhlenbeck algorithm. Selection pressure is controlled with a new feature of the evolutionary biology research. That is, the ratio of the variance of non-synonymous to synonymous selection coefficients (vN/vS). Overall, the package represents a promising addition to the phylogenetics modelling literature in at least two ways. First, it extends the capacity to generate molecular data to validate codon models of evolution. It offers an uncommon opportunity to thoroughly examine the implications of different fitness landscape settings and the interactions among different evolutionary factors. Second, it creates an exciting avenue for the development of sophisticated evolutionary models that are functions of the introduced vN/vS statistic.
-
  <!-- badges: start -->
-  [![Bioc history](https://bioconductor.org/shields/years-in-bioc/scoup.svg)](https://bioconductor.org/packages/release/bioc/html/scoup.html#since)
-  [![DOI](https://zenodo.org/badge/DOI/10.18129/B9.bioc.scoup.svg)](https://doi.org/10.18129/B9.bioc.scoup)
+  [![DOI](https://zenodo.org/badge/DOI/10.18129/B9.bioc.scoup.svg)](https://www.bioconductor.org/packages/devel/bioc/html/scoup.html)
   <!-- badges: end -->
+A R package intended for the Bioconductor platform. It avails an opportunity to simulate molecular data that are primarily useful for investigating the effects of Darwinian natural selection. Concepts from the population genomics and the phylogenetics literature are merged so that components of observed selection signatures in genetic sequences may be interrogated more thoroughly. This is achieved by combining (a.) the explicit use of selection coefficients by Halpern-Bruno in their popular 1998 mutation-selection codon model and (b.) the versatility of the stochastic Ornstein-Uhlenbeck algorithm. Selection pressure is controlled with a new feature of the evolutionary biology research. That is, the ratio of the variance of non-synonymous to synonymous selection coefficients (vN/vS). Overall, the package represents a promising addition to the phylogenetics modelling literature in at least two ways. First, it extends the capacity to generate molecular data to validate codon models of evolution. It offers an uncommon opportunity to thoroughly examine the implications of different fitness landscape settings and the interactions among different evolutionary factors. Second, it creates an exciting avenue for the development of sophisticated evolutionary models that are functions of the introduced vN/vS statistic.
 
 ## Contents
 - [CONTRIBUTING](CONTRIBUTING.md): information about how to contribute to the package and report bugs or seek assistance on how to use any of the functions in therein.
