### 2026-03-04 #9797 @thsadiq
@jromanowska just a follow-up to let you know that scoup 1.5.4 now reflects on [Bioconductor](https://bioconductor.org/packages/3.23/bioc/html/scoup.html)
