### 2026-03-05 #9797 @jromanowska
@thsadiq - just one small typo in the DOI, as [noted above](https://github.com/openjournals/joss-reviews/issues/9797#issuecomment-4003495760) - there is an unneeded 'x' in the end.
