### 2026-03-05 102104b
Corrected doi typo in paper.bib
