+++ paper/paper.bib
@@ -166,7 +166,7 @@
   volume  = {26},
   number  = {8},
   pages   = {1879--1888},
-  doi     = {10.1093/molbev/msp098x},
+  doi     = {10.1093/molbev/msp098},
 }
 
 @article{sipos2011,
