### 2026-03-05 #9797 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
