### 2026-03-15 #9797 @Kevin-Mattheus-Moerman
👉 please create a release over at your GitHub repository: https://github.com/thsadiq/scoup/tags. Next tag this release with the right version tag. Your DOI seems to say `Bioconductor version: Release (3.22)`, whereas this review issues says `1.5.4`. The three tags, over at GitHub, the archived version, and here, need to be to same. Let me know which tag we should use here.

👉 the title and author set do not match. If possible please update these.
👉 This may need updating depending on the other version/tag related issue highlighted.

- [x] Copyedit pass/checked for typos
👉 The paper currently uses both British English (see utilises, discretised, analysed, summarised) and American English (see parameterized, prioritized) please use one consistently
