### 2026-03-16 #9797 @thsadiq
Thank you @Kevin-Mattheus-Moerman.  My responses to the final checks are below.

👉 please create a release over at your GitHub repository: https://github.com/thsadiq/scoup/tags. Next tag this release with the right version tag. Your DOI seems to say Bioconductor version: Release (3.22), whereas this review issues says 1.5.4. The three tags, over at GitHub, the archived version, and here, need to be to same. Let me know which tag we should use here.
  - A [Github release tag](https://github.com/thsadiq/scoup/releases/tag/v1.5.4) with the appropriate v1.5.4 has been created as advised.
  - The Bioconductor DOI was an issue during the review. In hindsight, I should have generated a new DOI when I archived on Zenodo. The [developmental version of the package on Bioconductor](https://bioconductor.org/packages/devel/bioc/html/scoup.html) is 1.5.4. Bioconductor DOI however always links to the published version that could be about 6 months behind. To resolve this, I have removed the Bioconductor link from Zenodo by uploading a new version with Figshare. I wish to appeal that the package archive be  updated to the new DOI [**10.6084/m9.figshare.31743991**](https://doi.org/10.6084/m9.figshare.31743991). The Zenodo version has been updated to reflect this new DOI.

👉 the title and author set do not match. If possible please update these.
  - The newly created [Zenodo/Figshare](https://doi.org/10.6084/m9.figshare.31743991) archive eliminates this mismatch.

- [x] Archive listed version tag matches tagged release (and includes a potential v).
👉 This may need updating depending on the other version/tag related issue highlighted.
  - The version details associated with the updated [Zenodo/Figshare](https://doi.org/10.6084/m9.figshare.31743991) archives has been prefixed with a `v`.

- [x] Copyedit pass/checked for typos
👉 The paper currently uses both British English (see utilises, discretised, analysed, summarised) and American English (see parameterized, prioritized) please use one consistently.
  - Thanks for detecting this. Paper has been edited accordingly.

Please let me know if there are any other concerns you require me to resolve. Thanks for the efforts and hard work toward ensuring that the final outputs are flawless.
