+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: scoup
-Version: 1.5.4
-Date: 2026-02-03
+Version: 1.5.5
+Date: 2026-03-17
 License: GPL (>= 2)
 Authors@R: c(person("Hassan", "Sadiq", role=c("aut","cre","cph"),
                     email="hassan.t.sadiq@gmail.com",
