### 2026-03-17 #9797 @Kevin-Mattheus-Moerman
@openjournals/dev this DOI is still not resolving for me. Can you have a look? Should I re-accept?
