### 2026-03-17 #9797 @thsadiq
@Kevin-Mattheus-Moerman I believe it is a general problem and not a (re-)acceptance error. None of the seven recently accepted JOSS papers link to their assigned DOIs.
