### 2026-03-17 #9797 @xuanxu
It looks like the doi.org [servers](https://doi.statuspage.io/) are having problems: https://doi.statuspage.io/incidents/f4cqcppc1q41
